EMP-553: Refactored to move database related models to their own package
EMP-553: Refactored to move database related models to their own package
