+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/controller/Crm14Controller.java
@@ -9,7 +9,7 @@ import uk.gov.justice.laa.crime.equinity.historicaldata.config.CrmFormDetailsCri
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.api.Crm14InterfaceApi;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.Crm14FormDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.mapper.Crm14Mapper;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm14.Crm14AttachmentModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.Crm14AttachmentModel;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm14.Crm14Model;
 import uk.gov.justice.laa.crime.equinity.historicaldata.service.Crm14AttachmentService;
 import uk.gov.justice.laa.crime.equinity.historicaldata.service.CrmFileService;
