+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/AttachmentStoreRepository.java
@@ -2,7 +2,7 @@ package uk.gov.justice.laa.crime.equinity.historicaldata.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm14.Crm14AttachmentModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.Crm14AttachmentModel;
 
 import java.util.List;
