+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/CrmFormsViewRepository.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.equinity.historicaldata.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import uk.gov.justice.laa.crime.equinity.historicaldata.config.CrmFormModelInterface;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmFormViewModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.CrmFormViewModel;
 
 
 public interface CrmFormsViewRepository extends JpaRepository<CrmFormViewModel, Integer>, JpaSpecificationExecutor<CrmFormModelInterface> {
