+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/TaskImageFilesRepository.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.equinity.historicaldata.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmFormModelInterface;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.TaskImageFilesModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.TaskImageFilesModel;
 
 
 public interface TaskImageFilesRepository extends JpaRepository<TaskImageFilesModel, Long>, JpaSpecificationExecutor<CrmFormModelInterface> {}
