+++ equinity-historical-data/src/test/java/uk/gov/justice/laa/crime/equinity/historicaldata/service/SearchServiceTest.java
@@ -14,7 +14,7 @@ import uk.gov.justice.laa.crime.equinity.historicaldata.config.CrmFormSearchCrit
 import uk.gov.justice.laa.crime.equinity.historicaldata.exception.ResourceNotFoundException;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.SearchCrmFormDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.SearchResultDTO;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmFormViewModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.CrmFormViewModel;
 import uk.gov.justice.laa.crime.equinity.historicaldata.repository.CrmFormsViewRepository;
 
 @SpringBootTest
